+++ packages/mini_program_tooling/lib/src/mini_program_embedding_initializer.dart
@@ -518,6 +518,14 @@ class MiniPrograms {
     appId: '$_publicDemoAppId',
     title: '$_publicDemoTitle',
   );
+
+  static const values = <MiniProgramInfo>[
+    publicDemo,
+  ];
+
+  static const byAppId = <String, MiniProgramInfo>{
+    publicDemo.appId: publicDemo,
+  };
 }
 ''';
   }